Move Git and Helm testservers to dedicated pkgs
This is a better approach, because:

1) The Helm server pulls in a lot of (indirect) dependencies from Helm,
   while those are unnecessary dependencies for most of the projects.
2) The Git server has a replace statement in go.mod, given this needs
   to be copied over by any depending project, and most of the projects
   do not need to mock any git server, moving it to a dedicated pacakge
   clears them from this burden.
