+++ testserver/http.go
@@ -25,6 +25,8 @@ import (
 	"path/filepath"
 )
 
+// NewTempHTTPServer returns a HTTPServer with a newly created temp
+// dir as the docroot.
 func NewTempHTTPServer() (*HTTPServer, error) {
 	tmpDir, err := ioutil.TempDir("", "http-test-")
 	if err != nil {
@@ -34,6 +36,7 @@ func NewTempHTTPServer() (*HTTPServer, error) {
 	return srv, nil
 }
 
+// NewHTTPServer returns a HTTPServer with the given docroot set.
 func NewHTTPServer(docroot string) *HTTPServer {
 	root, err := filepath.Abs(docroot)
 	if err != nil {
@@ -44,17 +47,24 @@ func NewHTTPServer(docroot string) *HTTPServer {
 	}
 }
 
+// HTTPServer is an HTTP/S server for testing purposes.
+// It can serve files from the configured docroot and offers
+// a lightweight middleware configuration option.
 type HTTPServer struct {
 	docroot    string
 	middleware func(http.Handler) http.Handler
 	server     *httptest.Server
 }
 
+// WithMiddleware configures the middleware of the HTTPServer, this can for
+// example be used to configure Basic Auth credentials. It should be called
+// before starting the server, or requires a stop/start cycle.
 func (s *HTTPServer) WithMiddleware(m func(handler http.Handler) http.Handler) *HTTPServer {
 	s.middleware = m
 	return s
 }
 
+// Start starts the HTTPServer.
 func (s *HTTPServer) Start() {
 	s.server = httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		handler := http.FileServer(http.Dir(s.docroot))
@@ -66,7 +76,8 @@ func (s *HTTPServer) Start() {
 	}))
 }
 
-func (s *HTTPServer) StartTLS(cert, key, ca []byte) error {
+// StartTLS starts the TLS HTTPServer with the given TLS configuration.
+func (s *HTTPServer) StartTLS(cert, key, ca []byte, serverName string) error {
 	s.server = httptest.NewUnstartedServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		handler := http.FileServer(http.Dir(s.docroot))
 		if s.middleware != nil {
@@ -88,23 +99,27 @@ func (s *HTTPServer) StartTLS(cert, key, ca []byte) error {
 	cp.AppendCertsFromPEM(ca)
 	config.RootCAs = cp
 
-	config.ServerName = "example.com"
+	config.ServerName = serverName
 	s.server.TLS = &config
 
 	s.server.StartTLS()
 	return nil
 }
 
+// Stop stops the HTTPServer, if started.
 func (s *HTTPServer) Stop() {
 	if s.server != nil {
 		s.server.Close()
 	}
 }
 
+// Root returns the configured docroot of the HTTPServer.
 func (s *HTTPServer) Root() string {
 	return s.docroot
 }
 
+// URL returns the address the HTTPServer is listening at,
+// if started.
 func (s *HTTPServer) URL() string {
 	if s.server != nil {
 		return s.server.URL
